requirements.txt
streamlit
openai
python-dotenv
beautifulsoup4
requests
PyPDF2
google-generativeai
fpdf2